Fixed a few features in the Dba
Fixed the class list not updating after saving/adding and also made it
so you don't need to have a class to save/add.

There are still a few weird features that we need to address:

The combobox does this flicker thing sometimes and I'm not sure why.
Adding and updating DBAs sometimes gets weird errors too.